Fix unnecessary re-render on highScore update
Previously, every re-render of CardGrid caused all the Cards to be
reorded. Due to that, two state updates in quick succession would cause
a visual 'flicker' where the Cards get reorded twice. This version of
the code only reorders the Cards when a change to currentStreakIds
occurs, meaning the user has to make a guess, correct or wrong.